machine-learning-client/requirements.txt
tensorflow==2.12.0
flask
numpy
pillow
pymongo
onnxruntime
opencv-python-headless
scipy
